chore: moved coingecko types into coingecko folder
