+++ packages/swapper/package.json
@@ -27,7 +27,7 @@
     "axios": "^0.21.4",
     "bignumber.js": "^9.0.1",
     "retry-axios": "^2.6.0",
-    "web3": "1.6.0"
+    "web3": "^1.6.1"
   },
   "devDependencies": {
     "@shapeshiftoss/asset-service": "^1.6.1",
